moved WIP(work in progress) GPQ(grafana prometheus quickwit) logic into a library
